Added some features and fixed some bugs. Beta release candidate.

+++ FunnySounds/MainWindow.xaml.cs
@@ -1,18 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Net;
-using System.Text;
 using System.Threading.Tasks;
 using System.Timers;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Navigation;
 
 namespace FunnySounds
 {
@@ -21,7 +14,7 @@ namespace FunnySounds
     /// </summary>
     public partial class MainWindow : Window
     {
-        //Add renaming sounds in the future, Save user setting in some .json - Global odds / Per Sound odds OR Multiply local odds by Global Odds. add new XAML. randomize volume / chance to randomize volume?
+        //randomize volume / chance to randomize volume?
         private string dataDir = Structs.dataDir;
         private WebClient webClient = new WebClient();
         private Timer playerTimer = new Timer();
@@ -41,7 +34,7 @@ namespace FunnySounds
             if (_UserData != null)
             {
                 userData = _UserData;
-                globalFrequencySlider.Value = userData.globalFrequency;
+                globalFrequencySlider.Value = Math.Pow(userData.globalFrequency, (1 / 1.5f));
                 globalVolumeSlider.Value = userData.globalVolume;
             }
             else
@@ -65,6 +58,7 @@ namespace FunnySounds
 
         private async void FirstTimeSetup()
         {
+            //Some default Ultra Mega Funny sounds -_-
             var Sound1 = Utils.DownloadAudio("https://www.myinstants.com/media/sounds/fart-with-reverb.mp3", "Fard");
             var Sound2 = Utils.DownloadAudio("https://www.myinstants.com/media/sounds/yt1s_wU4BGgD.mp3", "WhatTheDogDoin");
             var Sound3 = Utils.DownloadAudio("https://www.myinstants.com/media/sounds/deeznuts_2.mp3", "DeezNuts");
@@ -73,11 +67,14 @@ namespace FunnySounds
 
             await Task.WhenAll(Sound1, Sound2, Sound3, Sound4, Sound5);
 
-            availableSounds.Add(Sound1.Result);
-            availableSounds.Add(Sound2.Result);
-            availableSounds.Add(Sound3.Result);
-            availableSounds.Add(Sound4.Result);
-            availableSounds.Add(Sound5.Result);
+            if (Sound1.Result != null && Sound2.Result != null && Sound3.Result != null && Sound4.Result != null && Sound5.Result != null) //Couldn't be bothered with making a statement for every one separately. I should have made an array of these but... I didnt 
+            {
+                availableSounds.Add(Sound1.Result);
+                availableSounds.Add(Sound2.Result);
+                availableSounds.Add(Sound3.Result);
+                availableSounds.Add(Sound4.Result);
+                availableSounds.Add(Sound5.Result);
+            }
 
             foreach (var sound in availableSounds.ToList())
             {
@@ -92,28 +89,6 @@ namespace FunnySounds
 
         public void LoadBasicAudio()
         {
-            //var Sound1 = Utils.DownloadAudio("https://www.myinstants.com/media/sounds/fart-with-reverb.mp3", "Fard");
-            //var Sound2 = Utils.DownloadAudio("https://www.myinstants.com/media/sounds/yt1s_wU4BGgD.mp3", "WhatTheDogDoin");
-            //var Sound3 = Utils.DownloadAudio("https://www.myinstants.com/media/sounds/deeznuts_2.mp3", "DeezNuts");
-            //var Sound4 = Utils.DownloadAudio("https://www.myinstants.com/media/sounds/among-us-roundstart.mp3", "Amugus");
-            //var Sound5 = Utils.DownloadAudio("https://www.myinstants.com/media/sounds/vine-boom.mp3", "Vine Boom");
-
-            //await Task.WhenAll(Sound1, Sound2, Sound3, Sound4, Sound5);
-
-            //availableSounds.Add(Sound1.Result);
-            //availableSounds.Add(Sound2.Result);
-            //availableSounds.Add(Sound3.Result);
-            //availableSounds.Add(Sound4.Result);
-            //availableSounds.Add(Sound5.Result);
-
-            //foreach (var fileSound in userData.sounds)
-            //{
-            //    if (!availableSounds.Any(x => x.name == fileSound.name))
-            //        availableSounds.Add(fileSound);
-            //    else
-            //        availableSounds[availableSounds.IndexOf(availableSounds.First(x => x.name == fileSound.name))] = fileSound;
-            //}
-
             availableSounds = userData.sounds;
 
             foreach (var sound in availableSounds.ToList())
@@ -129,22 +104,33 @@ namespace FunnySounds
             float _maxOdds = 0f;
             availableSounds.ForEach(x => { if (x.isEnabled) _maxOdds += x.odds; });
             float rngNum = (float)(rnd.NextDouble() * _maxOdds);
-            float _currentOdds = 0f;
+            float currentOdds = 0;
             foreach (Structs.Sound sound in availableSounds)
             {
-                _currentOdds += sound.odds;
-
-                if(_currentOdds > rngNum && sound != null && sound.isEnabled)
+                if (sound != null && sound.isEnabled)
                 {
-                    Utils.PlayAudio(sound.path, sound.volume * globalVolume);
-                    return;
+                    currentOdds += sound.odds;
+                    if (currentOdds > rngNum)
+                    {
+                        var player = Utils.PlayAudio(sound.path, sound.volume * globalVolume);
+                        if (player != null)
+                        {
+                            sound.mediaPlayers.Add(player);
+                            return;
+                        }
+                        else
+                        {
+                            MessageBox.Show($"There was an error while playing the sound: {sound.name}");
+                            return;
+                        }
+                    }
                 }
             }
         }
 
         private void GlobalFrequencyChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
         {
-            globalOdds = (float)e.NewValue;
+            globalOdds = (float)Math.Pow(e.NewValue, 1.5d);
             if (currentGlobalFrequency != null)
             {
                 currentGlobalFrequency.Content = $"{Math.Round(globalOdds * 100)}%";
@@ -170,6 +156,12 @@ namespace FunnySounds
             var link = linkTextBox.Text;
             var name = linkSoundNameBox.Text;
 
+            if (userData.sounds.Any(sound => sound.name == name))
+            {
+                MessageBox.Show("Sound with this name already exists");
+                return;
+            }
+
             if (name == null || link == null)
             {
                 MessageBox.Show("Please put in the name/link");
@@ -182,7 +174,7 @@ namespace FunnySounds
             }
 
             var sound = await Utils.DownloadAudio(link, name);
-            availableSounds.Add(sound);
+            if (sound == null) return;
             userData.sounds.Add(sound);
 
             var soundControl = new Controls.SoundControl(sound);
@@ -199,23 +191,37 @@ namespace FunnySounds
             Utils.SaveUserData(userData);
         }
 
-        public void RemoveSound(Structs.Sound soundToRemove)
+        public async void RemoveSound(Controls.SoundControl soundControl)
         {
-            userData.sounds.Remove(soundToRemove);
-
-            foreach (var child in soundsPanel.Children)
+            if (soundControl == null || soundControl.DataContext == null) return;
+            if (soundControl.DataContext is Structs.Sound soundData)
             {
-                var dataContext = (child as Controls.SoundControl).DataContext as Structs.Sound;
-                if (DataContext != null && dataContext.path == soundToRemove.path) //NULL
-                    soundsPanel.Children.Remove(child as UIElement);
+                userData.sounds.Remove(soundData);
+
+                soundData.isEnabled = false;
+                soundsPanel.Children.Remove(soundControl);
+
+                if (Path.GetDirectoryName(soundData.path) == Structs.dataDir)
+                {
+                    soundData.mediaPlayers.ForEach(p => p.Close());
+                    await Task.Delay(10); //Better safe than sorry
+                    File.Delete(soundData.path);
+                }
             }
+
+            //foreach (var child in soundsPanel.Children)
+            //{
+            //    var dataContext = (child as Controls.SoundControl).DataContext as Structs.Sound;
+            //    if (DataContext != null && dataContext.path == soundToRemove.path) //NULL
+            //        soundsPanel.Children.Remove(child as UIElement);
+            //}
         }
 
         private void FileDropped(object sender, DragEventArgs e)
         {
             var name = fileSoundNameBox.Text;
 
-            if(name == null)
+            if (name == null)
             {
                 MessageBox.Show("Please put in the name/link");
                 return;
@@ -233,6 +239,7 @@ namespace FunnySounds
                 droppedSoundFile = sound;
 
                 fileTextBox.Text = Path.GetFileName(file);
+
 
                 //Could use Foreach, but there is no good way to deal with naming all the files, and it's not that useful, so I'm gettng rid of it for now (for eternity);
 
@@ -262,38 +269,43 @@ namespace FunnySounds
         private void CopyDraggedFilesClicked(object sender, RoutedEventArgs e)
         {
             if (droppedSoundFile == null) return;
+            if (userData.sounds.Any(sound => sound.name == droppedSoundFile.name))
+            {
+                MessageBox.Show("Sound with this name already exists");
+                return;
+            }
             var newSound = droppedSoundFile;
-            var soundToReplace = availableSounds.First(x=> x.path == newSound.path);
             try
             {
                 File.Copy(droppedSoundFile.path, Path.Combine(Structs.dataDir, Path.GetFileName(droppedSoundFile.path)));
             }
-            catch (System.IO.IOException _ex)
+            catch (System.IO.IOException)
             {
                 var dialogResult = MessageBox.Show("There was an error while copying the file (File with that same name might already exist. Do You want to override the current file?)", "File Error", MessageBoxButton.YesNo);
-                if(dialogResult == MessageBoxResult.OK)
+                if (dialogResult == MessageBoxResult.OK)
                 {
                     try
                     {
                         File.Copy(droppedSoundFile.path, Path.Combine(Structs.dataDir, Path.GetFileName(droppedSoundFile.path)), true);
                     }
-                    catch (Exception ex)
+                    catch (Exception)
                     {
                         MessageBox.Show("The issue persists, it might be caused by the file, its localization, or my lack of programming skills");
                         return;
                     }
                 }
-                else if(dialogResult == MessageBoxResult.No)
+                else if (dialogResult == MessageBoxResult.No)
                 {
                     return;
                 }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 MessageBox.Show("There was an error while copying the file");
                 return;
             }
             newSound.path = Path.Combine(Structs.dataDir, Path.GetFileName(droppedSoundFile.path));
+            newSound.name = fileSoundNameBox.Text;
 
             //availableSounds.Add(newSound);
             //availableSounds.Remove(soundToReplace);
@@ -332,7 +344,13 @@ namespace FunnySounds
 
         private void LinkDraggedFilesClicked(object sender, RoutedEventArgs e)
         {
-            if(droppedSoundFile == null) return;
+            if (droppedSoundFile == null) return;
+            if (userData.sounds.Any(sound => sound.name == droppedSoundFile.name))
+            {
+                MessageBox.Show("Sound with this name already exists");
+                return;
+            }
+            droppedSoundFile.name = fileSoundNameBox.Text;
             //availableSounds.Add(droppedSoundFile);
             userData.sounds.Add(droppedSoundFile);
